requirements.txt
keyboard==0.13.5
PySDL2==0.9.12
